Add remap-stdint.rsp for Vulkan
Only uint64_t=ulong affects the generation, but the others are included to be slightly more complete.
This does not fix the enum backing types though, which also differ between Linux/Windows.

See this conversation in the Silk.NET Discord for more information: https://discord.com/channels/521092042781229087/1376331581198827520/1376628539536969949

Relevant ClangSharp issue: https://github.com/dotnet/ClangSharp/issues/574

+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureBuildSizesInfoKHR.gen.cs
@@ -41,7 +41,7 @@ public unsafe partial struct AccelerationStructureBuildSizesInfoKHR
             "VK_KHR_deferred_host_operations+VK_VERSION_1_2",
         ]
     )]
-    public nuint AccelerationStructureSize;
+    public ulong AccelerationStructureSize;
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile(
@@ -52,7 +52,7 @@ public unsafe partial struct AccelerationStructureBuildSizesInfoKHR
             "VK_KHR_deferred_host_operations+VK_VERSION_1_2",
         ]
     )]
-    public nuint UpdateScratchSize;
+    public ulong UpdateScratchSize;
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile(
@@ -63,5 +63,5 @@ public unsafe partial struct AccelerationStructureBuildSizesInfoKHR
             "VK_KHR_deferred_host_operations+VK_VERSION_1_2",
         ]
     )]
-    public nuint BuildScratchSize;
+    public ulong BuildScratchSize;
 }
