+++ sources/Vulkan/Vulkan/Vulkan/AccelerationStructureTrianglesOpacityMicromapEXT.gen.cs
@@ -60,7 +60,7 @@ public unsafe partial struct AccelerationStructureTrianglesOpacityMicromapEXT
             "VK_KHR_acceleration_structure+VK_VERSION_1_3",
         ]
     )]
-    public nuint IndexStride;
+    public ulong IndexStride;
 
     [NativeTypeName("uint32_t")]
     [SupportedApiProfile(
