+++ sources/Vulkan/Vulkan/Vulkan/BindImageMemoryInfo.gen.cs
@@ -48,5 +48,5 @@ public unsafe partial struct BindImageMemoryInfo
         ["VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.1"
     )]
-    public nuint MemoryOffset;
+    public ulong MemoryOffset;
 }
