+++ sources/Vulkan/Vulkan/Vulkan/BindVertexBufferIndirectCommandEXT.gen.cs
@@ -19,7 +19,7 @@ public partial struct BindVertexBufferIndirectCommandEXT
             "VK_VERSION_1_3",
         ]
     )]
-    public nuint BufferAddress;
+    public ulong BufferAddress;
 
     [NativeTypeName("uint32_t")]
     [SupportedApiProfile(
