+++ sources/Vulkan/Vulkan/Vulkan/BindVertexBufferIndirectCommandNV.gen.cs
@@ -16,7 +16,7 @@ public partial struct BindVertexBufferIndirectCommandNV
         ["VK_NV_device_generated_commands"],
         ImpliesSets = ["VK_VERSION_1_1+VK_KHR_buffer_device_address", "VK_VERSION_1_2"]
     )]
-    public nuint BufferAddress;
+    public ulong BufferAddress;
 
     [NativeTypeName("uint32_t")]
     [SupportedApiProfile(
