+++ sources/Vulkan/Vulkan/Vulkan/BufferCopy.gen.cs
@@ -17,7 +17,7 @@ public partial struct BufferCopy
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public nuint SrcOffset;
+    public ulong SrcOffset;
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile(
@@ -25,7 +25,7 @@ public partial struct BufferCopy
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public nuint DstOffset;
+    public ulong DstOffset;
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile(
@@ -33,5 +33,5 @@ public partial struct BufferCopy
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public nuint Size;
+    public ulong Size;
 }
