+++ sources/Vulkan/Vulkan/Vulkan/BufferCreateInfo.gen.cs
@@ -40,7 +40,7 @@ public unsafe partial struct BufferCreateInfo
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public nuint Size;
+    public ulong Size;
 
     [NativeTypeName("VkBufferUsageFlags")]
     [SupportedApiProfile(
