+++ sources/Vulkan/Vulkan/Vulkan/BufferImageCopy2.gen.cs
@@ -20,7 +20,7 @@ public unsafe partial struct BufferImageCopy2
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint BufferOffset;
+    public ulong BufferOffset;
 
     [NativeTypeName("uint32_t")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
