+++ sources/Vulkan/Vulkan/Vulkan/BufferMemoryBarrier2.gen.cs
@@ -20,19 +20,19 @@ public unsafe partial struct BufferMemoryBarrier2
 
     [NativeTypeName("VkPipelineStageFlags2")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint SrcStageMask;
+    public ulong SrcStageMask;
 
     [NativeTypeName("VkAccessFlags2")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint SrcAccessMask;
+    public ulong SrcAccessMask;
 
     [NativeTypeName("VkPipelineStageFlags2")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint DstStageMask;
+    public ulong DstStageMask;
 
     [NativeTypeName("VkAccessFlags2")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint DstAccessMask;
+    public ulong DstAccessMask;
 
     [NativeTypeName("uint32_t")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
@@ -48,9 +48,9 @@ public unsafe partial struct BufferMemoryBarrier2
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint Offset;
+    public ulong Offset;
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile("vulkan", ["VK_VERSION_1_3", "VK_VERSION_1_4"], MinVersion = "1.3")]
-    public nuint Size;
+    public ulong Size;
 }
