+++ sources/Vulkan/Vulkan/Vulkan/BufferViewCreateInfo.gen.cs
@@ -55,7 +55,7 @@ public unsafe partial struct BufferViewCreateInfo
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public nuint Offset;
+    public ulong Offset;
 
     [NativeTypeName("VkDeviceSize")]
     [SupportedApiProfile(
@@ -63,5 +63,5 @@ public unsafe partial struct BufferViewCreateInfo
         ["VK_VERSION_1_0", "VK_VERSION_1_1", "VK_VERSION_1_2", "VK_VERSION_1_3", "VK_VERSION_1_4"],
         MinVersion = "1.0"
     )]
-    public nuint Range;
+    public ulong Range;
 }
