+++ sources/Vulkan/Vulkan/Vulkan/ClusterAccelerationStructureBuildTriangleClusterInfoNV.gen.cs
@@ -149,7 +149,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint IndexBuffer;
+    public ulong IndexBuffer;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -157,7 +157,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint VertexBuffer;
+    public ulong VertexBuffer;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -165,7 +165,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint GeometryIndexAndFlagsBuffer;
+    public ulong GeometryIndexAndFlagsBuffer;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -173,7 +173,7 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint OpacityMicromapArray;
+    public ulong OpacityMicromapArray;
 
     [NativeTypeName("VkDeviceAddress")]
     [SupportedApiProfile(
@@ -181,5 +181,5 @@ public partial struct ClusterAccelerationStructureBuildTriangleClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint OpacityMicromapIndexBuffer;
+    public ulong OpacityMicromapIndexBuffer;
 }
