+++ sources/Vulkan/Vulkan/Vulkan/ClusterAccelerationStructureInstantiateClusterInfoNV.gen.cs
@@ -60,7 +60,7 @@ public partial struct ClusterAccelerationStructureInstantiateClusterInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint ClusterTemplateAddress;
+    public ulong ClusterTemplateAddress;
 
     [SupportedApiProfile(
         "vulkan",
