+++ sources/Vulkan/Vulkan/Vulkan/ClusterAccelerationStructureMoveObjectsInfoNV.gen.cs
@@ -17,5 +17,5 @@ public partial struct ClusterAccelerationStructureMoveObjectsInfoNV
         ["VK_NV_cluster_acceleration_structure"],
         ImpliesSets = ["VK_KHR_acceleration_structure"]
     )]
-    public nuint SrcAccelerationStructure;
+    public ulong SrcAccelerationStructure;
 }
