+++ sources/Vulkan/Vulkan/Vulkan/ConditionalRenderingBeginInfoEXT.gen.cs
@@ -52,7 +52,7 @@ public unsafe partial struct ConditionalRenderingBeginInfoEXT
             "VK_EXT_conditional_rendering+VK_VERSION_1_1",
         ]
     )]
-    public nuint Offset;
+    public ulong Offset;
 
     [NativeTypeName("VkConditionalRenderingFlagsEXT")]
     [SupportedApiProfile(
